PARQUET-400: Move SeekableInputStream to io package.
